test(extensions): split provider-neutral core from finance view tests
Core presentation tests and the lifecycle smoke now use a synthetic,
structure-only view and assert the compact `detail_ref` status contract,
including a budget regression that a ready item never inlines a view.
Manifest tests cover the generic `kind`/`view_schema` token rules. The
finance decision-research view tests move to a finance-owned test module.
Update the extensions reference to describe the provider-neutral core,
pluggable view validators, the compact status matrix, and public-safe
visibility.
