+++ tests/extensions/test_extension_presentation.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-from copy import deepcopy
 from datetime import datetime, timezone
 import json
 from pathlib import Path
@@ -13,7 +12,6 @@ from loopx.extensions.presentation import (
     collect_active_extension_presentation_surfaces,
     default_extension_projection_root,
     publish_extension_projection,
-    validate_decision_research_view,
 )
 from loopx.extensions.runtime import (
     disable_extension,
@@ -24,206 +22,27 @@ from loopx.extensions.runtime import (
 
 
 def _valid_view() -> dict[str, object]:
+    # Core keeps the `view` opaque unless an extension registers a validator for
+    # the declared view_schema. These lifecycle tests use a synthetic,
+    # structure-only view so they exercise generic Core behavior without
+    # depending on any one domain's schema. Finance owns and tests its own
+    # decision-research view in tests/extensions/test_finance_presentation_view.py.
     return {
-        "identity": {
-            "title": "Synthetic Technology Research",
-            "subtitle": "Evidence-gated decision review",
-            "as_of": "2026-01-15T12:00:00+00:00",
-            "evidence_cutoff": "2026-01-15",
-        },
-        "adjudication": {
-            "status": "insufficient_evidence",
-            "label": "Insufficient Evidence",
-            "summary": "The frozen gates do not yet support a selected conclusion.",
-            "confidence": "medium",
-        },
-        "metrics": [
+        "headline": "Synthetic lifecycle view",
+        "generated_label": "unchanged",
+        "sections": [
             {
-                "id": "validated-alpha",
-                "label": "Validated alpha",
-                "value": "0",
-                "detail": "No company-specific residual passed every frozen gate.",
-                "tone": "warning",
-            },
-            {
-                "id": "method-state",
-                "label": "Method state",
-                "value": "unchanged",
-                "detail": "The active method was not promoted or replaced.",
-                "tone": "neutral",
-            },
-        ],
-        "dashboard_summaries": [
-            {
-                "id": "adjudication-summary",
-                "label": "Current adjudication",
-                "title": "Evidence remains insufficient",
-                "summary": "Continue monitoring frozen event gates.",
-                "tone": "warning",
-                "destination_anchor": "executive-adjudication",
+                "id": "summary",
+                "title": "Summary",
+                "body": "Opaque passthrough content for the lifecycle proof.",
             }
         ],
-        "layers": [
-            {
-                "id": "beta",
-                "order": 1,
-                "label": "Beta",
-                "status": "supported",
-                "summary": "Discount-rate exposure explains part of the move.",
-                "evidence_points": [
-                    "Broad peer dispersion remains limited.",
-                    "Rate sensitivity is visible across the synthetic group.",
-                ],
-            },
-            {
-                "id": "residual-alpha",
-                "order": 4,
-                "label": "Residual alpha",
-                "status": "rejected",
-                "summary": "Validated company alpha = 0.",
-                "evidence_points": [
-                    "The candidate did not pass persistence controls.",
-                    "Counterevidence remains unresolved.",
-                ],
-            },
-        ],
-        "entities": [
-            {
-                "entity_id": "synthetic-cloud",
-                "symbol": "SYN",
-                "display_name": "Synthetic Cloud",
-                "classification": "Watchlist",
-                "status": "insufficient_evidence",
-                "confidence": "medium",
-                "inference": "A quality business is not yet a validated mispricing.",
-                "observations": [
-                    {
-                        "id": "observation-range",
-                        "label": "Observation range",
-                        "kind": "observation_range",
-                        "value": "90-100 synthetic units",
-                        "as_of": "2026-01-15T12:00:00+00:00",
-                        "source_ref": "filing:syn-q4",
-                        "source_type": "company_filing",
-                        "confidence": "high",
-                        "invalidation": "A verified close below 90 with worsening fundamentals.",
-                    }
-                ],
-                "scenario_estimates": [
-                    {
-                        "scenario": "bull",
-                        "label": "Bull scenario estimate",
-                        "value": "140 synthetic units",
-                        "horizon": "24 months",
-                        "probability": 0.25,
-                        "assumptions": [
-                            "Growth reaccelerates.",
-                            "Free cash flow conversion improves.",
-                        ],
-                    },
-                    {
-                        "scenario": "base",
-                        "label": "Base scenario estimate",
-                        "value": "112 synthetic units",
-                        "horizon": "24 months",
-                        "probability": 0.5,
-                        "assumptions": [
-                            "Growth remains durable.",
-                            "The valuation multiple is stable.",
-                        ],
-                    },
-                    {
-                        "scenario": "bear",
-                        "label": "Bear scenario estimate",
-                        "value": "72 synthetic units",
-                        "horizon": "24 months",
-                        "probability": 0.25,
-                        "assumptions": [
-                            "Growth slows.",
-                            "The valuation multiple compresses.",
-                        ],
-                    },
-                ],
-                "counterevidence": [
-                    "Capital intensity may remain above the frozen assumption."
-                ],
-                "thesis_breakers": [
-                    "Two consecutive periods of slowing growth and weaker cash conversion."
-                ],
-                "next_events": [
-                    "Next official earnings release.",
-                    "Updated capital expenditure guidance.",
-                ],
-            }
-        ],
-        "research_ledger": [
-            {
-                "case_id": "case-synthetic-cloud",
-                "label": "Synthetic Cloud residual-alpha case",
-                "gate_states": [
-                    {
-                        "gate_id": "de-beta",
-                        "label": "De-beta control",
-                        "status": "passed",
-                        "summary": "The peer control was available.",
-                    },
-                    {
-                        "gate_id": "persistence",
-                        "label": "Persistence",
-                        "status": "failed",
-                        "summary": "The residual did not persist.",
-                    },
-                ],
-                "decision": "rejected",
-                "summary": "The company-specific alpha claim was rejected.",
-                "evidence_refs": ["filing:syn-q4", "market:synthetic-peer-control"],
-            }
-        ],
-        "event_gates": [
-            {
-                "event_id": "E1",
-                "label": "Synthetic cloud earnings",
-                "status": "pending",
-                "observation_window": "Next official reporting window",
-                "frozen_hypothesis": "Returns depend on monetization, not capex alone.",
-                "observables": [
-                    "Cloud growth versus frozen guidance.",
-                    "Capital expenditure and free cash flow direction.",
-                ],
-                "current_evidence": [
-                    "No official result was available at the evidence cutoff."
-                ],
-                "supports": [
-                    "Growth above the frozen range with stable cash conversion."
-                ],
-                "refutes": [
-                    "Higher capital expenditure without measurable cloud return."
-                ],
-                "thesis_breakers": [
-                    "Official guidance shows deteriorating returns on investment."
-                ],
-                "next_review": "After the official filing is available.",
-            }
-        ],
-        "method_state": {
-            "revision": "candidate-v1",
-            "lifecycle_state": "active_method_unchanged",
-            "active_method_changed": False,
-            "summary": "Active method unchanged.",
-        },
-        "boundary": {
-            "research_aid_only": True,
-            "investment_advice": False,
-            "trading_allowed": False,
-            "raw_provider_payload_recorded": False,
-            "private_source_content_read": False,
-        },
     }
 
 
 def _provider_projection() -> dict[str, object]:
     return {
-        "schema_version": "finance_research_dashboard_packet_v0",
+        "schema_version": "synthetic_research_packet_v0",
         "presentation_projection": {
             "schema_version": "extension_presentation_projection_v0",
             "surface_id": "investment-research",
@@ -237,7 +56,7 @@ def _provider_projection() -> dict[str, object]:
                 "supersedes": [],
                 "superseded_by": None,
             },
-            "view_schema": "decision_research_dashboard_v0",
+            "view_schema": "synthetic_dashboard_v0",
             "view": _valid_view(),
         },
     }
@@ -296,10 +115,10 @@ timeout_seconds = 5
 
 [[presentation_surfaces]]
 id = "investment-research"
-kind = "decision_research_dashboard"
+kind = "synthetic_dashboard"
 title = "Investment Research"
-view_schema = "decision_research_dashboard_v0"
-visibility = "owner-only"
+view_schema = "synthetic_dashboard_v0"
+visibility = "public-safe"
 empty_state_title = "No validated research yet"
 empty_state_detail = "Publish a validated projection."
 """,
@@ -326,187 +145,6 @@ def _installed_projection_extension(
     state_file = tmp_path / "runtime" / "extensions" / "state.json"
     installed = install_extension(manifest, state_file=state_file, execute=True)
     return state_file, installed
-
-
-def test_decision_research_view_preserves_strict_research_truth() -> None:
-    validated = validate_decision_research_view(_valid_view())
-
-    assert validated["adjudication"]["status"] == "insufficient_evidence"
-    assert validated["metrics"][0]["value"] == "0"
-    assert validated["metrics"][1]["value"] == "unchanged"
-    assert validated["layers"][1]["status"] == "rejected"
-    assert validated["research_ledger"][0]["decision"] == "rejected"
-    assert [item["probability"] for item in validated["entities"][0]["scenario_estimates"]] == [
-        0.25,
-        0.5,
-        0.25,
-    ]
-    assert validated["boundary"]["trading_allowed"] is False
-
-
-@pytest.mark.parametrize(
-    ("mutator", "message"),
-    [
-        (
-            lambda view: view.update({"unsupported": "field"}),
-            "unsupported keys",
-        ),
-        (
-            lambda view: view["identity"].update({"as_of": "not-a-timestamp"}),
-            "as_of",
-        ),
-        (
-            lambda view: view["entities"][0]["observations"][0].update(
-                {"invalidation": ""}
-            ),
-            "invalidation",
-        ),
-        (
-            lambda view: view["entities"][0]["scenario_estimates"][0].update(
-                {"assumptions": []}
-            ),
-            "assumptions",
-        ),
-        (
-            lambda view: view["entities"][0]["scenario_estimates"][0].update(
-                {"probability": 0.4}
-            ),
-            "sum to 1",
-        ),
-        (
-            lambda view: view["entities"][0].update(
-                {"counterevidence": []}
-            ),
-            "counterevidence",
-        ),
-        (
-            lambda view: view["entities"][0].update(
-                {"thesis_breakers": []}
-            ),
-            "thesis_breakers",
-        ),
-        (
-            lambda view: view["identity"].update(
-                {"subtitle": "<script>alert(1)</script>"}
-            ),
-            "plain text",
-        ),
-        (
-            lambda view: view["identity"].update(
-                {"subtitle": "Owner account ID 998877 remains linked."}
-            ),
-            "sensitive material",
-        ),
-        (
-            lambda view: view["identity"].update(
-                {"subtitle": "Owner account_id 998877 remains linked."}
-            ),
-            "sensitive material",
-        ),
-        (
-            lambda view: view["identity"].update(
-                {"subtitle": "Research source: .codex/goals/private-research.md"}
-            ),
-            "local path",
-        ),
-        (
-            lambda view: view["identity"].update(
-                {"subtitle": "Research source: ../.codex/goals/private-research.md"}
-            ),
-            "local path",
-        ),
-        (
-            lambda view: view["identity"].update(
-                {"subtitle": "Research source: project/.local/private-research.md"}
-            ),
-            "local path",
-        ),
-        (
-            lambda view: view["layers"][0]["evidence_points"].append(
-                "/tmp/private-research.json"
-            ),
-            "local path",
-        ),
-        (
-            lambda view: view["layers"][0].update(
-                {"raw_provider_response": "secret"}
-            ),
-            "forbidden key",
-        ),
-        (
-            lambda view: view["boundary"].update({"investment_advice": True}),
-            "research boundary",
-        ),
-        (
-            lambda view: view.update(
-                {
-                    "dashboard_summaries": [
-                        *view["dashboard_summaries"],
-                        *deepcopy(view["dashboard_summaries"]),
-                        *deepcopy(view["dashboard_summaries"]),
-                        *deepcopy(view["dashboard_summaries"]),
-                    ]
-                }
-            ),
-            "at most 3",
-        ),
-    ],
-)
-def test_decision_research_view_rejects_invalid_or_private_content(
-    mutator,
-    message: str,
-) -> None:
-    view = deepcopy(_valid_view())
-    mutator(view)
-
-    with pytest.raises(ValueError, match=message):
-        validate_decision_research_view(view)
-
-
-def test_decision_research_view_rejects_non_finite_probability() -> None:
-    view = deepcopy(_valid_view())
-    view["entities"][0]["scenario_estimates"][0]["probability"] = float("nan")
-
-    with pytest.raises(ValueError, match="finite"):
-        validate_decision_research_view(view)
-
-
-def test_decision_research_view_allows_explicit_public_evidence_url() -> None:
-    view = _valid_view()
-    view["research_ledger"][0]["evidence_refs"].append(
-        "https://www.sec.gov/Archives/edgar/data/000000/example.htm"
-    )
-
-    validated = validate_decision_research_view(view)
-
-    assert validated["research_ledger"][0]["evidence_refs"][-1].startswith(
-        "https://www.sec.gov/"
-    )
-
-
-@pytest.mark.parametrize(
-    "reference",
-    [
-        "http://example.com/evidence",
-        "https://localhost/evidence",
-        "https://127.0.0.1/evidence",
-        "https://private.example.internal/evidence",
-        "https://example.com/private/report?"
-        + "".join(("to", "ken"))
-        + "=synthetic-value",
-        "https://example.com/report?"
-        + "".join(("refresh_", "to", "ken"))
-        + "=synthetic-value",
-    ],
-)
-def test_decision_research_view_rejects_unsafe_evidence_urls(
-    reference: str,
-) -> None:
-    view = _valid_view()
-    view["research_ledger"][0]["evidence_refs"].append(reference)
-
-    with pytest.raises(ValueError, match="evidence reference"):
-        validate_decision_research_view(view)
 
 
 def test_projection_publication_dry_run_does_not_invoke_or_write(
@@ -568,8 +206,8 @@ def test_projection_publication_binds_revision_hash_and_exact_readback(
     assert persisted["extension_id"] == "test-research-extension"
     assert persisted["extension_revision"] == installed["revision"]
     assert persisted["surface_id"] == "investment-research"
-    assert persisted["visibility"] == "owner-only"
-    assert persisted["view"]["metrics"][0]["value"] == "0"
+    assert persisted["visibility"] == "public-safe"
+    assert persisted["view"]["headline"] == "Synthetic lifecycle view"
     assert projection_file.stat().st_mode & 0o777 == 0o600
     assert not list(projection_file.parent.glob(".*.tmp.*"))
 
@@ -593,9 +231,9 @@ def test_failed_projection_publication_preserves_previous_file(
     previous_bytes = projection_file.read_bytes()
 
     invalid = _provider_projection()
-    invalid["presentation_projection"]["view"]["entities"][0][
-        "scenario_estimates"
-    ][0]["probability"] = 0.9
+    invalid["presentation_projection"]["view"]["sections"][0][
+        "api_key"
+    ] = "forbidden-secret"
     _projection_provider(tmp_path / "provider", projection=invalid)
     doctor_installed_extension(
         "test-research-extension",
@@ -603,7 +241,7 @@ def test_failed_projection_publication_preserves_previous_file(
         execute=True,
     )
 
-    with pytest.raises(ValueError, match="sum to 1"):
+    with pytest.raises(ValueError, match="forbidden key"):
         publish_extension_projection(
             "test-research-extension",
             "investment-research",
@@ -634,7 +272,7 @@ def test_projection_publication_rejects_undeclared_surface(
 def test_active_presentation_surfaces_project_empty_ready_and_review_due(
     tmp_path: Path,
 ) -> None:
-    state_file, _ = _installed_projection_extension(tmp_path)
+    state_file, installed = _installed_projection_extension(tmp_path)
 
     empty = collect_active_extension_presentation_surfaces(
         state_file=state_file,
@@ -644,6 +282,7 @@ def test_active_presentation_surfaces_project_empty_ready_and_review_due(
     assert empty["empty_count"] == 1
     assert empty["items"][0]["state"] == "empty"
     assert "view" not in empty["items"][0]
+    assert "detail_ref" not in empty["items"][0]
 
     publish_extension_projection(
         "test-research-extension",
@@ -652,13 +291,27 @@ def test_active_presentation_surfaces_project_empty_ready_and_review_due(
         request={"schema_version": "synthetic_request_v0"},
         execute=True,
     )
+    projection_file = (
+        default_extension_projection_root(state_file)
+        / "test-research-extension"
+        / "investment-research.json"
+    )
+    persisted = json.loads(projection_file.read_text(encoding="utf-8"))
     ready = collect_active_extension_presentation_surfaces(
         state_file=state_file,
         now=datetime(2026, 1, 20, tzinfo=timezone.utc),
     )
     assert ready["ready_count"] == 1
     assert ready["items"][0]["state"] == "ready"
-    assert ready["items"][0]["view"]["metrics"][0]["value"] == "0"
+    # Status is a compact index: it carries a content-addressed pointer to the
+    # published projection, never the full (provider-owned) view inline.
+    assert "view" not in ready["items"][0]
+    assert ready["items"][0]["detail_ref"] == {
+        "extension_id": "test-research-extension",
+        "surface_id": "investment-research",
+        "extension_revision": installed["revision"],
+        "payload_sha256": persisted["payload_sha256"],
+    }
 
     review_due = collect_active_extension_presentation_surfaces(
         state_file=state_file,
@@ -666,9 +319,10 @@ def test_active_presentation_surfaces_project_empty_ready_and_review_due(
     )
     assert review_due["review_due_count"] == 1
     assert review_due["items"][0]["state"] == "review_due"
-    assert review_due["items"][0]["view"]["method_state"][
-        "active_method_changed"
-    ] is False
+    assert "view" not in review_due["items"][0]
+    assert review_due["items"][0]["detail_ref"]["payload_sha256"] == (
+        persisted["payload_sha256"]
+    )
 
 
 def test_active_presentation_surfaces_hide_disabled_or_stale_extension(
@@ -725,7 +379,7 @@ def test_active_presentation_surfaces_fail_closed_on_current_file_corruption(
         / "investment-research.json"
     )
     persisted = json.loads(projection_file.read_text(encoding="utf-8"))
-    persisted["view"]["metrics"][0]["value"] = "forged"
+    persisted["view"]["headline"] = "forged"
     projection_file.write_text(json.dumps(persisted), encoding="utf-8")
 
     collection = collect_active_extension_presentation_surfaces(
@@ -736,6 +390,7 @@ def test_active_presentation_surfaces_fail_closed_on_current_file_corruption(
     assert collection["items"][0]["state"] == "invalid"
     assert collection["items"][0]["diagnostic"] == "projection_hash_invalid"
     assert "view" not in collection["items"][0]
+    assert "detail_ref" not in collection["items"][0]
 
 
 def test_active_presentation_surfaces_do_not_reuse_old_revision_projection(
